restructure shared components
