+++ src/widgets/shared/AutoComplete/spec.js
