+++ src/widgets/shared/Icon/index.js
@@ -1,5 +1,19 @@
-import Icon from './Icon.react';
+import React from 'react';
 
-export {
-  Icon,
-};
+import prefix from '../../../utils/style';
+import classNames from 'classnames';
+
+const { icon } = prefix(['icon'], 'Icon');
+
+function iconClass(iconId) {
+  return classNames(
+          prefix([iconId], 'Icon')[iconId],
+          icon
+        );
+}
+
+const Icon = (props) => (
+  <span className={iconClass(props.id)}></span>
+);
+
+export default Icon;
