+++ src/widgets/shared/List/index.css
@@ -1,0 +1,3 @@
+.list {
+
+}
