+++ src/widgets/shared/ListItem/index.css
@@ -1,7 +1,3 @@
-.list {
-
-}
-
 .listItem {
   border-bottom: 1px solid #eee;
   background-color: #fff;
