+++ src/widgets/shared/list/index.js
@@ -1,6 +1,16 @@
-import List from './List.react';
-import ListItem from './ListItem.react';
-export {
-  List,
-  ListItem,
+import React from 'react';
+import prefix from '../../../utils/style';
+
+const { list } = prefix(['list'], 'List');
+
+const List = (props) => (
+  <div className={list}>
+    {props.children}
+  </div>
+);
+
+List.propTypes = {
+  children: React.PropTypes.node,
 };
+
+export default List;
