--- src/widgets/shared/panel/PanelContent.react.js
@@ -1,16 +1,0 @@
-import React from 'react';
-
-const PanelContent = (props) => {
-  let ContentElement = props.children;
-  return (
-    <div>
-      {ContentElement}
-    </div>
-  );
-};
-
-PanelContent.propTypes = {
-  children: React.PropTypes.node,
-};
-
-export default PanelContent;
