--- src/widgets/shared/panel/PanelFooter.react.js
@@ -1,16 +1,0 @@
-import React from 'react';
-
-const PanelFooter = (props) => {
-  let ContentElement = props.children;
-  return (
-    <div>
-      {ContentElement}
-    </div>
-  );
-};
-
-PanelFooter.propTypes = {
-  children: React.PropTypes.node,
-};
-
-export default PanelFooter;
