--- src/widgets/shared/panel/index.js
@@ -1,9 +1,0 @@
-import PanelHeader from './PanelHeader.react';
-import PanelContent from './PanelContent.react';
-import PanelFooter from './PanelFooter.react';
-
-export {
-  PanelHeader,
-  PanelContent,
-  PanelFooter,
-};
