+++ src/widgets/webphone/DialPad/index.js
@@ -3,9 +3,8 @@ import classNames from 'classnames';
 
 import LPN from 'google-libphonenumber';
 
-import { PanelHeader, PanelContent, PanelFooter } from '../../shared/Panel/';
-import { Input } from '../../shared/AutoComplete/';
-import { Icon } from '../../shared/Icon';
+import Input from '../../shared/AutoComplete/';
+import Icon from '../../shared/Icon';
 
 import Dialer from '../Dialer';
 import CallerBar from '../CallerBar';
@@ -111,7 +110,7 @@ export default class DialPad extends React.PureComponent {
             getString={this.props.getString}
           />
         </div>
-        <PanelContent>
+        <div>
           <div>
             <Input
               className={phoneInput}
@@ -123,8 +122,8 @@ export default class DialPad extends React.PureComponent {
               <Dialer handleClick={this.boundHandleClick} />
             </div>
           </div>
-        </PanelContent>
-        <PanelFooter>
+        </div>
+        <div>
           <div className={line}>
             <button
               className={callButton}
@@ -133,7 +132,7 @@ export default class DialPad extends React.PureComponent {
               <Icon id={'icon-uniAE'} />
             </button>
           </div>
-        </PanelFooter>
+        </div>
       </div>
     );
   }
